+++ Track/ViewModel/ChalaniTobillVM.cs
@@ -6,6 +6,6 @@ namespace Track.ViewModel
     public class ChalaniTobillVM
     {
         public BillClass Bill { get; set; }
-        public List<int> Rates { get; set; }
+        public List<decimal> Rates { get; set; }
     }
 }
